pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "comtypes>=1.4.9",
    "docx2pdf>=0.1.8",
    "docx>=0.2.4",
    "docxtpl>=0.19.1",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "reportlab>=4.2.5",
    "werkzeug>=3.1.3",
    "pillow>=11.1.0",
]
